requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
boto==2.46.1
Babel==2.4.0
BeautifulSoup==3.2.1
cached-property==1.3.0
cffi==1.10.0
cryptography==1.8.1
defusedxml==0.5.0
Django==1.8.18
django-crequest==2016.3.16
django-haystack==2.6.0
django-markdown-deux==1.0.5
django-mptt==0.8.7
django-recaptcha==1.2.1
django-registration==2.2
django-storages==1.5.2
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
isodate==0.5.4
lxml==3.7.3
magnitude==0.9.4
Markdown==2.6.8
markdown2==2.3.3
olefile==0.44
packaging==16.8
paramiko==2.1.2
Pillow==4.1.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
pysolr==3.6.0
python-memcached==1.58
pytz==2017.2
recaptcha-client==1.0.6
requests==2.13.0
requests-toolbelt==0.7.1
rollbar==0.13.11
six==1.10.0
sorl-thumbnail==12.3
uWSGI==2.0.15
zeep==1.4.1
